Add first version of the PWA
A user can click a button to start the microphone real-time playback. If you view the web development console, you will see that frequency data is collected and output to the console log in real time. This data can be used to visualise the audio from the microphone (which is yet to be implemented). This works offline as a progressive web app.